requirements/common.txt
asgiref==3.4.1
Django==3.1.14
pytz==2021.3
sqlparse==0.4.2
pillow
djangorestframework
markdown
django-filter
djangorestframework-jwt
django-pydenticon
python-dotenv